feat: Reorganize project structure to segregate app modules
Project structure has been reorganized for clearer readability and maintainability. This change includes adjustments in directory structure to properly categorize application modules ('core', 'store') under the 'candleaf' folder. This update improves the overall organization of the project and eases the process of navigating through different modules. Changed references in code files to reflect new structure. Performing this for better modularity and separation of concerns in our Django project. Other alterations include changing 'candleaf' to 'config' for settings module to make it clearer.
