+++ backend/candleaf/dev/products_data/images/Fragrant Cinnamon.png
