+++ backend/candleaf/dev/products_data/images/Pink Sugar Crystals.png
