+++ backend/candleaf/dev/products_data/images/Spiced Mint.png
