+++ backend/candleaf/dev/products_data/images/Sweet Strawberry.png
