+++ backend/candleaf/store/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class StoreConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'store'
+    name = 'candleaf.store'
