+++ backend/config/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for candleaf project.
+Django settings for config project.
 
 Generated by 'django-admin startproject' using Django 4.1.7.
 
@@ -54,8 +54,8 @@ INSTALLED_APPS = [
     'django_filters',
     'djoser',
     'social_django',
-    'core',
-    'store'
+    'candleaf.core',
+    'candleaf.store'
 ]
 
 MIDDLEWARE = [
@@ -71,7 +71,7 @@ MIDDLEWARE = [
     'django.middleware.common.CommonMiddleware'
 ]
 
-ROOT_URLCONF = 'candleaf.urls'
+ROOT_URLCONF = 'config.urls'
 
 TEMPLATES = [
     {
@@ -90,7 +90,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'candleaf.wsgi.application'
+WSGI_APPLICATION = 'config.wsgi.application'
 
 # Database
 # https://docs.djangoproject.com/en/4.1/ref/settings/#databases
@@ -136,7 +136,7 @@ USE_TZ = True
 
 STATIC_URL = 'static/'
 MEDIA_URL = 'media/'
-MEDIA_ROOT = os.path.join(BASE_DIR, 'media')
+MEDIA_ROOT = os.path.join(BASE_DIR, 'candleaf', 'media')
 
 # Default primary key field type
 # https://docs.djangoproject.com/en/4.1/ref/settings/#default-auto-field
